rename files to eips
