+++ crates/primitives/src/transaction/eip2930.rs
@@ -1,9 +1,7 @@
-use crate::{Bytes, ChainId, TransactionKind};
-use std::mem;
-
-use reth_codecs::{main_codec, Compact};
-
 use super::access_list::AccessList;
+use crate::{Bytes, ChainId, TransactionKind};
+use reth_codecs::{main_codec, Compact};
+use std::mem;
 
 /// Transaction with an [`AccessList`] ([EIP-2930](https://eips.ethereum.org/EIPS/eip-2930)).
 #[main_codec]
